pyproject.toml
[project]
name = "market-translator"
version = "0.1.0"
description = "Marketplace Schema Translator + Rate-Limit Agent (FastAPI + LangGraph + DSPy)"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langgraph>=0.2.12",
    "dspy-ai>=2.5.0",
    "numpy",
    "pandas",
    "pyyaml",
    "tenacity",
    "rich",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
